Extract a CompetitorQuery
